refactor: reshuffle module types
