+++ src/ducks/actions/reward.ts
@@ -9,7 +9,7 @@ export const processRewards = (
   if (playerOutcome.skills) {
     for (const skill of Object.keys(playerOutcome.skills)) {
       state.playerCharacter.skillScores[skill].value +=
-        playerOutcome.skills[skill].amountGained;
+        playerOutcome.skills[skill].amountGained || 0;
     }
   }
 };
