+++ src/module-types/decision.ts
@@ -1,5 +1,16 @@
-import { ModuleOption } from "./option";
 import { ModuleScene, SceneType } from "./scene";
+import { ModuleOutcome } from "./outcome";
+
+export type ModuleOnSelection = {
+  displayText?: string;
+  onSuccess: ModuleOutcome;
+};
+
+export type ModuleOption = {
+  displayText: string;
+  onSelection: ModuleOnSelection;
+  skillRequirements?: Record<string, number>;
+};
 
 export type ModuleDecision = ModuleScene & {
   type: SceneType.DECISION;
