--- src/module-types/skill.ts
@@ -1,3 +1,0 @@
-export type ModuleSkillDescriptions = Record<string, string>;
-
-export type ModuleSkillScore = { value: number };
